Lots of updates to roles.yaml.
Change the layout to have each VM run the CM server and agent and not to have a
central CM server. (memory and ports)
Tweak the OS postinstalls to better work with Cloudera's expectations of
software repo availability, SElinux, and UTF8 support. (shell provisioners)
Make the puppetmaster read the modules we have prepared for it. (shell
provisioners)